+++ src/config.ts
@@ -17,7 +17,7 @@ export const FEE_RATE_DIVIDER = 10_000;
 
 const goerli: Config = {
   networkId: NetworkEnum.GOERLI,
-  subgraphUrl: 'https://api.thegraph.com/subgraphs/name/quent043/badass-graph-yamakasi-mothafuc\n',
+  subgraphUrl: 'https://api.thegraph.com/subgraphs/name/quent043/badass-graph-yamakasi-mothafuc',
   // subgraphUrl: 'https://api.thegraph.com/subgraphs/name/talentlayer/talent-layer-protocol',
   contracts: {
     talentLayerId: '0x11119eD887aeC1302e2cAF49942F891667A31BBc',
@@ -90,7 +90,8 @@ const fuji: Config = {
 
 const mumbai: Config = {
   networkId: NetworkEnum.MUMBAI,
-  subgraphUrl: 'https://api.thegraph.com/subgraphs/name/talentlayer/talent-layer-mumbai',
+  // subgraphUrl: 'https://api.thegraph.com/subgraphs/name/talentlayer/talent-layer-mumbai',
+  subgraphUrl: 'https://api.thegraph.com/subgraphs/name/quent043/badass-graph-yamakasi-mothafuc',
   contracts: {
     talentLayerId: '0x3F87289e6Ec2D05C32d8A74CCfb30773fF549306',
     serviceRegistry: '0x27ED516dC1df64b4c1517A64aa2Bb72a434a5A6D',
