--- src/modules/Kleros/components/DisputeDetail.tsx
@@ -1,9 +1,0 @@
-function DisputeItem() {
-  return (
-    <div className={'flex flex-row gap-2 rounded-xl p-4 border border-gray-200'}>
-      <div className={''}></div>
-    </div>
-  );
-}
-
-export default DisputeItem;
